dbuf: refactor dbuf_dirty into a set of functions.
The original function consists of a set of operations that apply to a
disjunct set of scenarios.  For the purpose of understanding what it
does to a particular type of buffer, it's difficult to follow.  Rectify
this by breaking it up.

This change attempts to have no functional impact, although some
operations may occur slightly differently.